+++ Snake.py
@@ -1,0 +1,68 @@
+import pygame
+from pygame.locals import *
+import time
+import random
+from pygamepopup.menu_manager import MenuManager
+
+SIZE = 40
+BACKGROUND_COLOR = (110, 110, 5)
+
+
+
+class Snake:
+    def __init__(self, parent_screen):
+        self.parent_screen = parent_screen
+        self.image = pygame.image.load("block.jpg").convert()
+        self.direction = 'down'
+
+        # starting the snake game with the snake having 20 blocks for length
+        self.length = 20
+        self.x = [40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40]
+        self.y = [40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40, 40]
+
+    def move_left(self):
+        self.direction = 'left'
+
+    def move_right(self):
+        self.direction = 'right'
+
+    def move_up(self):
+        self.direction = 'up'
+
+    def move_down(self):
+        self.direction = 'down'
+
+    def walk(self):
+        # update body
+        for i in range(self.length - 1, 0, -1):
+            self.x[i] = self.x[i - 1]
+            self.y[i] = self.y[i - 1]
+
+        # update head
+        if self.direction == 'left':
+            self.x[0] -= SIZE
+        if self.direction == 'right':
+            self.x[0] += SIZE
+        if self.direction == 'up':
+            self.y[0] -= SIZE
+        if self.direction == 'down':
+            self.y[0] += SIZE
+
+        self.draw()
+
+    def draw(self):
+        self.parent_screen.fill(BACKGROUND_COLOR)
+
+        for i in range(self.length):
+            self.parent_screen.blit(self.image, (self.x[i], self.y[i]))
+        pygame.display.flip()
+
+    def increase_length(self):
+        self.length += 1
+        self.x.append(-1)
+        self.y.append(-1)
+
+    def decrease_length(self):
+        self.length -= 1
+        self.x.append(-1)
+        self.y.append(-1)
